# GH-38728: [Go] ipc: put lz4 decompression buffers back into sync.Pool