@asubiotto 2023-11-15
It doesn't look like there are any existing benchmarks. I've profiled our production workload before/after this change and the sync.Pool `New` allocations do not show up anymore (here is the before profile I linked in the issue: https://pprof.me/d5122d09b9dae81844c3068f75e31b76/?profileType=profile%3Aalloc_space%3Abytes%3Aspace%3Abytes)