@zeroshade 2023-11-15
I wonder if it's possible to add a unit test or benchmark or otherwise to check and prevent this issue from recurring? This looks good to me in general. If there's a good way to reasonably add a test or other check then let's do that, but if not then I'm happy with this. Let me know what you think @asubiotto