@asubiotto 2023-11-16
```
pkg: github.com/apache/arrow/go/v15/arrow/ipc
                    │  benchmain   │             benchclose              │
                    │    sec/op    │    sec/op     vs base               │
Reader/codec=lz4-12   38.05µ ± 12%   12.97µ ± 19%  -65.92% (p=0.002 n=6)

                    │   benchmain   │             benchclose              │
                    │     B/op      │     B/op      vs base               │
Reader/codec=lz4-12   266.58Ki ± 0%   10.49Ki ± 1%  -96.07% (p=0.002 n=6)

                    │ benchmain  │            benchclose            │
                    │ allocs/op  │ allocs/op   vs base              │
Reader/codec=lz4-12   155.0 ± 0%   149.0 ± 0%  -3.87% (p=0.002 n=6)
```

I think we might be suffering from the same issue in writers btw, will take a look.