@asubiotto 2023-11-16
Added a commit with the benchmark and another commit to have writers release buffers. It's not noticeable in the benchmark because we're probably writing too little but it doesn't hurt:
```
goos: darwin
goarch: arm64
pkg: github.com/apache/arrow/go/v15/arrow/ipc
                          │  benchmain   │             benchclose              │
                          │    sec/op    │   sec/op     vs base                │
IPC/Writer/codec=plain-12   8.370µ ±  5%   8.264µ ± 2%   -1.25% (p=0.043 n=10)
IPC/Reader/codec=plain-12   11.41µ ±  4%   11.13µ ± 4%   -2.43% (p=0.014 n=10)
IPC/Writer/codec=zstd-12    1.089m ± 12%   1.096m ± 8%        ~ (p=0.912 n=10)
IPC/Reader/codec=zstd-12    38.22µ ±  3%   36.45µ ± 6%        ~ (p=0.052 n=10)
IPC/Writer/codec=lz4-12     16.09µ ±  2%   15.99µ ± 5%        ~ (p=0.436 n=10)
IPC/Reader/codec=lz4-12     40.79µ ±  4%   13.06µ ± 4%  -67.98% (p=0.000 n=10)
geomean                     37.10µ         30.26µ       -18.44%

                          │   benchmain   │               benchclose               │
                          │     B/op      │     B/op      vs base                  │
IPC/Writer/codec=plain-12    8.394Ki ± 0%   8.394Ki ± 0%        ~ (p=1.000 n=10) ¹
IPC/Reader/codec=plain-12    8.646Ki ± 0%   8.646Ki ± 0%        ~ (p=1.000 n=10) ¹
IPC/Writer/codec=zstd-12     35.52Mi ± 0%   35.52Mi ± 0%        ~ (p=0.810 n=10)
IPC/Reader/codec=zstd-12     25.63Ki ± 0%   25.63Ki ± 0%        ~ (p=0.063 n=10)
IPC/Writer/codec=lz4-12      14.21Ki ± 1%   14.16Ki ± 0%   -0.35% (p=0.043 n=10)
IPC/Reader/codec=lz4-12     266.53Ki ± 0%   10.55Ki ± 0%  -96.04% (p=0.000 n=10)
geomean                      79.69Ki        46.50Ki       -41.66%
¹ all samples are equal

                          │ benchmain  │             benchclose              │
                          │ allocs/op  │ allocs/op   vs base                 │
IPC/Writer/codec=plain-12   113.0 ± 0%   113.0 ± 0%       ~ (p=1.000 n=10) ¹
IPC/Reader/codec=plain-12   126.0 ± 0%   126.0 ± 0%       ~ (p=1.000 n=10) ¹
IPC/Writer/codec=zstd-12    271.0 ± 0%   271.0 ± 0%       ~ (p=1.000 n=10) ¹
IPC/Reader/codec=zstd-12    219.0 ± 0%   219.0 ± 0%       ~ (p=1.000 n=10) ¹
IPC/Writer/codec=lz4-12     167.0 ± 0%   167.0 ± 0%       ~ (p=1.000 n=10) ¹
IPC/Reader/codec=lz4-12     154.0 ± 0%   149.0 ± 0%  -3.25% (p=0.000 n=10)
geomean                     167.1        166.1       -0.55%
¹ all samples are equal
```