@zeroshade 2023-11-16
@asubiotto Looks like you've got a nil de-reference issue in the current version, i'm guessing there's a situation where you are calling close even if we didn't actually perform compression / decompression? Haven't looked into it myself yet