@asubiotto 2023-11-23
Apologies for the delay, I was on PTO. The issue was that I was double-closing the writer since I was unaware the `compress` function already closes the writer (`Close` is part of the interface via `io.WriterCloser`). I removed the last commit (writing only) since it was unnecessary. Should be good now.