@mapleFU 2023-11-23
I'm not so familiar with Arrow GoLang impl, just curious should we poolize the zstd decompress context ( or maybe it's already implemented ?)