@asubiotto 2023-11-27
I don't think the context itself is that big of an object and at least I haven't seen it be a limitation for my use case. The overhead here is mostly from the codec buffers. I would probably hold off unless we see a clear need.